+++ inc/classes/class-load-scripts.php
@@ -19,91 +19,91 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 	final class IPR_Load_Scripts {
 
 		/**
-		 * admin scripts 
+		 * Admin scripts 
 		 *
 		 * @var array $admin
 		 */
 		private $admin = [];
 
 		/**
-		 * core scripts for deregistration
+		 * Scripts for deregistration
 		 *
 		 * @var array $undo
 		 */
 		private $undo = [];
 
 		/**
-		 * core scripts
+		 * Core scripts
 		 *
 		 * @var array $core
 		 */
 		private $core = [];
 
 		/**
-		 * external scripts
+		 * External scripts
 		 *
 		 * @var array $external
 		 */
 		private $external = [];
 
 		/**
-		 * header scripts
+		 * Header scripts
 		 *
 		 * @var array $header
 		 */
 		private $header = [];
 
 		/**
-		 * footer scripts
+		 * Footer scripts
 		 *
 		 * @var array $footer
 		 */
 		private $footer = [];
 
 		/**
-		 * plugin scripts
+		 * Plugin scripts
 		 *
 		 * @var array $plugins
 		 */
 		private $plugins = [];
 
 		/**
-		 * page scripts
+		 * Page scripts
 		 *
 		 * @var array $page
 		 */
 		private $page = [];
 
 		/**
-		 * conditional scripts
+		 * Conditional scripts
 		 *
 		 * @var array $conditional
 		 */
 		private $conditional = [];
 
 		/**
-		 * front page scripts
+		 * Front page scripts
 		 *
 		 * @var array $front
 		 */
 		private $front = [];
 
 		/**
-		 * custom scripts
+		 * Custom scripts
 		 *
 		 * @var array $custom
 		 */
 		private $custom = [];
 
 		/**
-		 * login scripts
+		 * Login scripts
 		 *
 		 * @var array $login
 		 */
 		private $login = [];
 
 		/**
-		 * localize scripts
+		 * Localize scripts
 		 *
 		 * @var array $local
 		 */
@@ -124,20 +124,26 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 			// Front end only
 			if ( is_admin() ) { return; }
 
+			// Load parent scripts
+			add_action( 'wp_enqueue_scripts', 		[ $this, 'load_parent_scripts' ], 10 ); 
+
 			// Load scripts
-			add_action( 'wp_enqueue_scripts', 	[ $this, 'load_scripts' ] ); 
+			add_action( 'wp_enqueue_scripts', 		[ $this, 'load_child_scripts' ], 25 ); 
+
+			// Dequeueue scripts
+			add_action( 'wp_enqueue_scripts', 		[ $this, 'undo_scripts' ], 99 ); 
 
 			// Conditional header scripts
-			add_action( 'wp_enqueue_scripts', 	[ $this, 'conditional_scripts' ] ); 
+			add_action( 'wp_enqueue_scripts', 		[ $this, 'conditional_scripts' ] ); 
 
 			// Inline header scripts 
-			add_action( 'wp_head', 				[ $this, 'header_scripts' ], 99 );
+			add_action( 'wp_head', 					[ $this, 'header_scripts' ], 99 );
 
 			// Footer Scripts
-			add_action( 'wp_footer', 			[ $this, 'footer_scripts' ], 99 );
+			add_action( 'wp_footer', 				[ $this, 'footer_scripts' ], 99 );
 
 			// Analytics
-			add_action( 'wp_head', 				[ $this, 'analytics_script' ], 100 );
+			add_action( 'wp_head', 					[ $this, 'analytics_script' ], 100 );
 		}
 
 		/**
@@ -150,7 +156,7 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 			// Admin scripts: [ 'label' => [ 'hook', 'src', (array)deps, 'ver' ] ... ]
 			$this->admin = $this->set_key( $scripts, 'admin' );
 
-			// Core scripts for deregistration: [ 'script-name', 'script-name2' ... ]
+			// Core scripts for deregistration: [ 'script-name', [ 'script-name2', 'template' ] ... ]
 			$this->undo = $this->set_key( $scripts, 'undo' );
 
 			// Core scripts: [ 'script-name', 'script-name2' ... ]
@@ -229,15 +235,33 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 		//----------------------------------------------
 
 		/**
+		 * Load core, header & footer parent scripts 
+		 */
+		public function load_parent_scripts() { 
+
+			// Register & enqueue jquery core script as default
+			wp_enqueue_script( 'jquery' );
+
+			// Add base footer scripts
+			wp_register_script( 'ipress', IPRESS_JS_URL . '/theme.js', [ 'jquery' ], null, true ); 
+
+			// Set up core localization
+			$trans = [ 
+				'home_url' => home_url(), 
+				'ajax_url' => admin_url( 'admin-ajax.php' ),
+				'rest_url' => rest_url( '/' ) 
+			];
+			wp_localize_script( 'ipress', 'theme', $trans ); 
+
+			// Enqueue main script
+			wp_enqueue_script( 'ipress' );
+		}
+
+		/**
 		 * Load core, header & footer scripts 
 		 */
-		public function load_scripts() { 
+		public function load_child_scripts() { 
 	 
-			// Dequeue core scripts - restrict admin for compatibility
-			if ( ! is_admin() ) {
-				foreach ( $this->undo as $k=>$v ) { wp_deregister_script( $v ); }
-			}
-
 			// Register & enqueue core scripts
 			foreach ( $this->core as $k=>$v ) { wp_enqueue_script( $v ); }
 
@@ -364,6 +388,27 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 			// Validate & Localize
 			if ( isset( $h['name'] ) && isset( $h['trans'] ) )  { 
 				wp_localize_script( $key, $h['name'], $h['trans'] ); 
+			}
+		}
+
+		/**
+		 * Dequeue scripts 
+		 */
+		public function undo_scripts() { 
+	 
+			// Dequeue core scripts 
+			foreach ( $this->undo as $s ) { 
+
+				// Page template or global
+				if ( is_array( $s ) ) {
+					if ( is_page_template( $s[1] ) ) {
+						wp_dequeue_script( $s[0] );
+						wp_deregister_script( $s[0] ); 
+					}
+				} else {
+					wp_dequeue_script( $s );
+					wp_deregister_script( $s ); 
+				}
 			}
 		}