+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # Changelog
 * (07 August 2018). 1.0.0 Initial Parent Theme Fork 
+* (09 August 2018). 1.0.1 Refactor script & style dequeue
+* (13 August 2018). 1.0.2 Refactor config file